Fix a few things on new design
