+++ packages/coinstac-ui/app/render/components/maps/maps-csv-field.jsx
@@ -58,7 +58,7 @@ class MapsCsvField extends React.Component {
     this.state = {
       remainingHeader: [],
       selectedFiles: [],
-      autoMapError: true,
+      autoMapError: false,
     };
 
     this.autoMap = this.autoMap.bind(this);
